upgrades people (back to vs2012)
zlib to 1.3.2
boost to 1.75.0
qt to 5.5.1

- return back to webkit
- finally got username thing to work
- fix https bug
- fix authenticationhelper not using new json format for getcurrentuser

+++ App/v8datamodel/TeleportService.cpp
@@ -442,7 +442,7 @@ void TeleportService::TeleportThreadImpl(shared_ptr<const Reflection::ValueTable
                 StandardOut::singleton()->printf(MESSAGE_ERROR, "Teleport exception: %s", e.what());
             }
 
-            double timeToSleep = 1000.f - retryTimer.delta().msec();
+            int timeToSleep = (int)(1000.f - retryTimer.delta().msec());
             if (timeToSleep > 0.f)
             {
                 boost::this_thread::sleep(boost::posix_time::milliseconds(timeToSleep));
